refactoring - move methods between *Utils classes so that they are where they belong, and create new BeanUtils class
