+++ camel-idea-plugin/src/main/java/com/github/cameltooling/idea/Constants.java
@@ -1,0 +1,32 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.cameltooling.idea;
+
+public final class Constants {
+
+    public static final String OSGI_BLUEPRINT_NAMESPACE = "http://www.osgi.org/xmlns/blueprint/v1.0.0";
+    public static final String[] ACCEPTED_NAMESPACES = new String[]{
+        "http://camel.apache.org/schema/spring",
+        "http://camel.apache.org/schema/blueprint",
+        "http://www.springframework.org/schema/beans",
+        OSGI_BLUEPRINT_NAMESPACE
+    };
+
+    private Constants() {
+    }
+
+}
