+++ camel-idea-plugin/src/main/java/com/github/cameltooling/idea/util/CamelIdeaUtils.java
@@ -63,91 +63,6 @@ public final class CamelIdeaUtils implements Disposable {
         return ServiceManager.getService(CamelIdeaUtils.class);
     }
 
-    public boolean isBeanDeclaration(@NotNull PsiElement element) {
-        return enabledExtensions.stream().anyMatch(e -> e.isBeanDeclaration(element));
-    }
-
-    public boolean isPartOfCamelContext(@NotNull PsiElement element) {
-        return enabledExtensions.stream().anyMatch(e -> e.isPartOfCamelContext(element));
-    }
-
-    public Optional<ReferenceableBeanId> findReferenceableBeanId(Module module, String id) {
-        return enabledExtensions.stream()
-            .map(e -> e.findReferenceableBeanId(module, id))
-            .filter(Objects::nonNull)
-            .findAny();
-    }
-
-    public List<ReferenceableBeanId> findReferenceableBeanIds(Module module, Predicate<String> idCondition) {
-        return enabledExtensions.stream()
-            .map(e -> e.findReferenceableBeanIds(module, idCondition))
-            .flatMap(List::stream)
-            .collect(Collectors.toList());
-    }
-
-    public List<ReferenceableBeanId> findReferenceableBeanIdsByType(Module module, PsiType expectedBeanType) {
-        return findReferenceableBeanIdsByType(module, id -> true, expectedBeanType);
-    }
-
-    public List<ReferenceableBeanId> findReferenceableBeanIdsByType(Module module,
-                                                                    Predicate<String> idCondition,
-                                                                    PsiType expectedBeanType) {
-        PsiElementFactory elementFactory = JavaPsiFacade.getElementFactory(module.getProject());
-        return findReferenceableBeanIds(module, idCondition).stream()
-                .filter(ref -> {
-                    PsiClass psiClass = resolveToPsiClass(ref);
-                    if (psiClass != null) {
-                        PsiClassType beanType = elementFactory.createType(psiClass);
-                        return expectedBeanType.isAssignableFrom(beanType);
-                    }
-                    return false;
-                })
-                .collect(Collectors.toList());
-    }
-
-    public Optional<PsiClass> findReferencedBeanClass(PsiElement element) {
-        return findBeanReference(element)
-            .map(this::resolveToPsiClass);
-    }
-
-    private PsiClass resolveToPsiClass(ReferenceableBeanId ref) {
-        return (PsiClass) Optional.of(ref)
-                .map(ReferenceableBeanId::getReferencedClass)
-                .map(ReferencedClass::getReference)
-                .map(PsiReference::resolve)
-                .filter(e -> e instanceof PsiClass)
-                .orElse(null);
-    }
-
-    private Optional<ReferenceableBeanId> findBeanReference(PsiElement element) {
-        PsiReference[] references = element.getReferences();
-        return Arrays.stream(references)
-            .filter(r -> r instanceof BeanReference)
-            .map(r -> ((BeanReference) r).findReferenceableBeanId().orElse(null))
-            .filter(Objects::nonNull)
-            .findAny();
-    }
-
-    public Optional<PsiType> findExpectedBeanTypeAt(PsiElement location) {
-        return enabledExtensions.stream()
-                .map(e -> e.findExpectedBeanTypeAt(location))
-                .filter(Objects::nonNull)
-                .findAny();
-    }
-
-    public PsiClass getPropertyBeanClass(XmlTag propertyTag) {
-        XmlTag beanTag = propertyTag.getParentTag();
-        if (beanTag != null && CamelIdeaUtils.getService().isBeanDeclaration(beanTag)) {
-            IdeaUtils ideaUtils = IdeaUtils.getService();
-            return ideaUtils.findAttributeValue(beanTag, "class")
-                .map(ideaUtils::findClassReference)
-                .map(Optional::get)
-                .map(ideaUtils::resolveJavaClassReference)
-                .orElse(null);
-        }
-        return null;
-    }
-
     /**
      * Is the given element from the start of a Camel route, eg <tt>from</tt>, ot &lt;from&gt;.
      */
