+++ camel-idea-plugin/src/main/java/com/github/cameltooling/idea/util/JavaClassUtils.java
@@ -17,11 +17,16 @@
 package com.github.cameltooling.idea.util;
 
 import java.beans.Introspector;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
+
 import com.intellij.lang.jvm.annotation.JvmAnnotationAttribute;
 import com.intellij.openapi.Disposable;
+import com.intellij.openapi.components.ServiceManager;
 import com.intellij.openapi.project.Project;
 import com.intellij.psi.JavaPsiFacade;
 import com.intellij.psi.PsiAnnotation;
@@ -32,15 +37,21 @@ import com.intellij.psi.PsiJavaCodeReferenceElement;
 import com.intellij.psi.PsiReference;
 import com.intellij.psi.PsiType;
 import com.intellij.psi.impl.source.PsiClassReferenceType;
+import com.intellij.psi.impl.source.resolve.reference.impl.providers.JavaClassReference;
 import com.intellij.psi.search.GlobalSearchScope;
 import com.intellij.psi.search.ProjectScope;
 import com.intellij.psi.search.searches.AnnotatedElementsSearch;
 import com.intellij.psi.util.PsiUtil;
 import com.intellij.util.Query;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 
 public class JavaClassUtils implements Disposable {
+
+    public static JavaClassUtils getService() {
+        return ServiceManager.getService(JavaClassUtils.class);
+    }
 
     /**
      * @return the bean name of the {@link PsiClass} by check if the class is annotated with Component, Service, Repository or defaulting
@@ -90,25 +101,42 @@ public class JavaClassUtils implements Disposable {
         return Collections.emptyList();
     }
 
+    public JavaClassReference findClassReference(@NotNull PsiElement element) {
+        List<JavaClassReference> references = Arrays.stream(element.getReferences())
+                .filter(r -> r instanceof JavaClassReference)
+                .map(r -> (JavaClassReference) r)
+                .collect(Collectors.toList());
+        if (!references.isEmpty()) {
+            return references.get(references.size() - 1);
+        }
+        return null;
+    }
+
+    public PsiClass resolveClassReference(@NotNull PsiReference reference) {
+        final PsiElement resolveElement = reference.resolve();
+
+        if (resolveElement instanceof PsiClass) {
+            return (PsiClass) resolveElement;
+        } else if (resolveElement instanceof PsiField) {
+            final PsiType psiType = PsiUtil.getTypeByPsiElement(resolveElement);
+            if (psiType != null) {
+                return ((PsiClassReferenceType) psiType).resolve();
+            }
+        }
+
+        return null;
+    }
+
     /**
-     * @param referenceElement - Psi Code reference to resolve
+     * @param referenceElement - Psi Code reference element to resolve
      * @return Resolving the Psi reference and return the PsiClass
      */
-    public PsiClass resolveClassReference(PsiJavaCodeReferenceElement referenceElement) {
+    public PsiClass resolveClassReference(@Nullable PsiJavaCodeReferenceElement referenceElement) {
         if (referenceElement != null) {
             PsiReference reference = referenceElement.getReference();
 
             if (reference != null) {
-                final PsiElement resolveElement = reference.resolve();
-
-                if (resolveElement instanceof PsiClass) {
-                    return (PsiClass) resolveElement;
-                } else if (resolveElement instanceof PsiField) {
-                    final PsiType psiType = PsiUtil.getTypeByPsiElement(resolveElement);
-                    if (psiType != null) {
-                        return ((PsiClassReferenceType) psiType).resolve();
-                    }
-                }
+                return resolveClassReference(reference);
             }
         }
         return null;
